testcasegenerator/Antigravity Project One/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
python-dotenv==1.0.1
openai==1.50.0
requests==2.32.3
python-docx==1.1.2
google-generativeai==0.8.6
PyPDF2==3.0.1
pydantic[email]
pydantic-settings
faker
tenacity
